Add agency and car management features, including validation and detail views

+++ crud/delete_voiture.php
@@ -1,8 +1,27 @@
 <?php
 
+session_start();
+
 require_once '../config/db.php';
 
 $id = $_GET['id'];
+
+$sql = "SELECT * FROM voitures
+        WHERE id_voiture = ?
+        AND id_agence = ?";
+
+$stmt = $pdo->prepare($sql);
+
+$stmt->execute([
+    $id,
+    $_SESSION['id_agence']
+]);
+
+$voiture = $stmt->fetch(PDO::FETCH_ASSOC);
+
+if (!$voiture) {
+    die("Accès refusé");
+}
 
 $sql = "DELETE FROM images_voitures
         WHERE id_voiture = ?";
